refactor: General staticmethods moved to bmodel. Name change in base class (Domain model no db model)
